src/client/pyproject.toml
[tool.poetry]
name = "client"
version = "0.1.0"
description = "Shakesapp client"
authors = ["Yoshi Yamaguchi <yoshifumi@google.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.8"
grpcio = "^1.36.1"
protobuf = "^3.15.5"
grpcio-health-checking = "^1.36.1"
Flask = "^1.1.2"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.4"
isort = "^5.7.0"
grpcio-tools = "^1.36.1"

[tool.black]
extend-exclude = '''
^/(
    shakesapp_pb2_grpc.py
  | shakesapp_pb2.py
)
'''

[tool.poetry.scripts]
app = 'client:run'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
